Deduplicate inline `get_poc_diff` to `src/env.rs`.
